magipop-map-backend/requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
bitarray==2.8.1
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
Events==0.5
exceptiongroup==1.1.3
fastapi==0.103.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lz4==4.3.2
multidict==6.0.4
neo-mamba==2.0.2
neo3crypto==0.4
netaddr==0.8.0
orjson==3.9.5
psycopg2==2.9.7
pybiginteger==1.3.0
pybiginteger-stubs==1.3.0
pycares==4.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
referencing==0.30.2
rpds-py==0.10.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
yarl==1.9.2
